refactor(sync): split concurrent vec tests for GH727
