+++ specs/GH727/tasks.md
@@ -300,24 +300,30 @@ GH-727 / #727
 - [x] `SP727-T287` Owner: coordinator. Done when: `src/core/observability/metrics.rs` keeps production metrics structs, collector methods, and export/send behavior, then delegates tests with `#[path = "metrics_tests.rs"] mod tests;`. Verify: `tail -n 20 src/core/observability/metrics.rs`.
 - [x] `SP727-T288` Owner: coordinator. Done when: original inline metrics tests move to `src/core/observability/metrics_tests.rs` without assertion changes. Verify: `rg -n "test_prometheus_metrics_default|test_metrics_collector_with_datadog|test_record_request_updates_all_primary_metrics|test_record_cache_mixed|test_update_provider_health_overwrite|test_export_prometheus_with_health|test_send_to_datadog_with_client|test_record_request_duration|test_record_request_zero_tokens" src/core/observability/metrics_tests.rs`.
 - [x] `SP727-T289` Owner: verification owner. Done when: both touched observability metrics files are below U-16's 800-line ceiling and focused observability metrics tests pass. Verify: `wc -l src/core/observability/metrics.rs src/core/observability/metrics_tests.rs`; `cargo test core::observability::metrics --lib --all-features` passed 33 tests.
-- [ ] `SP727-T290` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the observability metrics tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
+- [x] `SP727-T290` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the observability metrics tranche. Verify: #873 PR body, green PR CI, GraphQL reviewThreads totalCount 0, merge commit `2cc5bedd1c16`, #727 reopened, and #727 update comment.
+- [x] `SP727-T291` Owner: coordinator. Done when: after the observability metrics tranche merges, the next #727 tranche is selected from the remaining queue with fresh tracked-file line-count evidence. Verify: at `origin/main@2cc5bedd`, `src/utils/sync/concurrent_vec.rs` is 807 lines and 3 tracked Rust files remain over the U-16 ceiling.
+- [x] `SP727-T292` Owner: coordinator. Done when: the ConcurrentVec tranche documents why this slice is a test extraction rather than a production sync utility split. Verify: `git diff -- specs/GH727/product.md specs/GH727/tech.md`.
+- [x] `SP727-T293` Owner: coordinator. Done when: `src/utils/sync/concurrent_vec.rs` keeps production `ConcurrentVec<T>` storage, methods, and trait impls, then delegates tests with `#[path = "concurrent_vec_tests.rs"] mod tests;`. Verify: `tail -n 20 src/utils/sync/concurrent_vec.rs`.
+- [x] `SP727-T294` Owner: coordinator. Done when: original inline ConcurrentVec tests move to `src/utils/sync/concurrent_vec_tests.rs` without assertion changes. Verify: `rg -n "test_new_and_default|test_push_and_pop|test_get_and_set|test_insert_and_remove|test_swap_remove|test_for_each_mut|test_contains_and_position|test_clone|test_concurrent_push|test_concurrent_read_write" src/utils/sync/concurrent_vec_tests.rs`.
+- [x] `SP727-T295` Owner: verification owner. Done when: both touched ConcurrentVec files are below U-16's 800-line ceiling and focused ConcurrentVec tests pass. Verify: `wc -l src/utils/sync/concurrent_vec.rs src/utils/sync/concurrent_vec_tests.rs`; `cargo test utils::sync::concurrent_vec --lib --all-features` passed 20 tests.
+- [ ] `SP727-T296` Owner: verification owner. Done when: formatting, SpecRail, diff check, lib all-features check, locked all-features check, default check, PR CI, and review-thread gate pass for the ConcurrentVec tranche. Verify: `cargo fmt --all -- --check`; `git diff --check`; `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`; `cargo check --lib --all-features`; `cargo check --all-features --locked`; `cargo check`; GitHub PR CI and review-thread query.
 
 ## 并行拆分
 
-This is a serial writable lane for the observability metrics file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for the ConcurrentVec file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/core/observability/metrics.rs`
-- `src/core/observability/metrics_tests.rs`
+- `src/utils/sync/concurrent_vec.rs`
+- `src/utils/sync/concurrent_vec_tests.rs`
 
 ## 验证
 
 - SpecRail packet review.
 - `cargo fmt --all -- --check`
 - `git diff --check`
-- `cargo test core::observability::metrics --lib --all-features`
+- `cargo test utils::sync::concurrent_vec --lib --all-features`
 - `cargo check --lib --all-features`
 - `cargo check --all-features --locked`
 - `cargo check`
@@ -325,5 +331,5 @@ Writable ownership for this lane:
 
 ## Handoff Notes
 
-This PR is the next #727 observability metrics unit-test extraction tranche and should use `Refs #727`, not `Closes #727`.
+This PR is the next #727 ConcurrentVec unit-test extraction tranche and should use `Refs #727`, not `Closes #727`.
 The issue should remain open until the final scan shows no Rust files over the U-16 ceiling.
